Add a test for multiple require()-loaded files.
Currently not working for the webpack 1 tests and I'm not sure why.

This seems to work in actual usage but for some reason the test runner
setup isn't creating the secondary output files.

I assume I'm setting something up incorrectly in the programmatic
webpack configuration or something else like that but I can't see what
it is at the moment.